Correctifs protection accès administrateurs et ajout style page authentification

+++ backend/controllers/post.js
@@ -5,19 +5,21 @@ const {
     Post,
     Comment,
     Like
-} = require('../models/index')
+} = require('../models/index');
+const checkAdminRights = require('../utils/checkAdminRights');
+const checkAuthorId = require('../utils/checkAuthorId');
+
 
 exports.getAllPosts = (req, res) => {
     Post.findAll({
-            include: [
-                {
-                    model : User
+            include: [{
+                    model: User
                 },
                 {
-                    model : Comment
+                    model: Comment
                 },
                 {
-                    model : Like
+                    model: Like
                 }
             ]
         })
@@ -32,6 +34,7 @@ exports.getAllPosts = (req, res) => {
 };
 
 exports.createPost = (req, res) => {
+    const isRequestFromUser = checkAuthorId(req);
     User.findOne({
             where: {
                 id: req.body.authorId
@@ -43,26 +46,32 @@ exports.createPost = (req, res) => {
                     error: "This user doesn't exist"
                 })
             }
-            Post.create({
-                    UserId: user.id,
-                    creationDate: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
-                    updateDate: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
-                    title: req.body.title,
-                    description: req.body.description,
-                    URLimage: req.file ? `${req.protocol}://${req.get("host")}/images/${req.file.filename}` : `${req.protocol}://${req.get("host")}/images/pikachu.jpg`
-                })
-                .then((post) => {
-                    res.status(200).json({
-                        message: "Post created",
-                        post
+            if (isRequestFromUser == user.id) {
+                Post.create({
+                        UserId: user.id,
+                        creationDate: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
+                        updateDate: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
+                        title: req.body.title,
+                        description: req.body.description,
+                        URLimage: req.file ? `${req.protocol}://${req.get("host")}/images/${req.file.filename}` : `${req.protocol}://${req.get("host")}/images/pikachu.jpg`
                     })
-                })
-                .catch((error) => {
-                    res.status(404).json({
-                        error: "Post not created",
-                        error
+                    .then((post) => {
+                        res.status(200).json({
+                            message: "Post created",
+                            post
+                        })
                     })
+                    .catch((error) => {
+                        res.status(404).json({
+                            error: "Post not created",
+                            error
+                        })
+                    })
+            } else {
+                return res.status(401).json({
+                    error
                 })
+            }
         })
         .catch((error) => {
             res.status(500).json({
@@ -76,8 +85,8 @@ exports.getOnePost = (req, res) => {
             where: {
                 id: req.params.id
             },
-            include: { 
-                model : User
+            include: {
+                model: User
             }
         })
         .then((post) => {
@@ -97,6 +106,7 @@ exports.getOnePost = (req, res) => {
 };
 
 exports.modifyPost = (req, res) => {
+    const isRequestFromUser = checkAuthorId(req);
     const postObject = req.file ? {
         ...req.body,
         URLimage: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
@@ -114,23 +124,28 @@ exports.modifyPost = (req, res) => {
                     error: "This post doesn't exist"
                 })
             }
-            foundPost.update({
-                    updateDate: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
-                    description: postObject.description,
-                    URLimage: postObject.URLimage
-                })
-                .then((updatedPost) => {
-                    res.status(200).json({
-                        updatedPost,
-                        message: 'This post has been updated'
+            if (isRequestFromUser == foundPost.UserId) {
+                foundPost.update({
+                        updateDate: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
+                        ...postObject
                     })
-                })
-                .catch((error) => {
-                    res.status(400).json({
-                        error,
-                        error: 'Could not update post'
+                    .then((updatedPost) => {
+                        res.status(200).json({
+                            updatedPost,
+                            message: 'This post has been updated'
+                        })
                     })
+                    .catch((error) => {
+                        res.status(400).json({
+                            error,
+                            error: 'Could not update post'
+                        })
+                    })
+            } else {
+                return res.status(401).json({
+                    error
                 })
+            }
         })
         .catch((error) => {
             res.status(500).json({
@@ -141,26 +156,32 @@ exports.modifyPost = (req, res) => {
 }
 
 exports.deletePost = (req, res) => {
+    const isRequestFromAdmin = checkAdminRights(req);
+    const isRequestFromUser = checkAuthorId(req);
     Post.findOne({
             where: {
                 id: req.params.id
             }
         })
         .then((foundPost) => {
-
             if (!foundPost) {
                 return res.status(404).json({
                     error: "This post doesn't exist"
                 })
             }
-            const filename = foundPost.URLimage.split("/images/")[1];
-            fs.unlink(`images/${filename}`, () => {
-                foundPost.destroy();
-                return res.status(200).json({
-                    message: "Post successfuly deleted"
+            if (isRequestFromUser == foundPost.UserId || isRequestFromAdmin === true) {
+                const filename = foundPost.URLimage.split("/images/")[1];
+                fs.unlink(`images/${filename}`, () => {
+                    foundPost.destroy();
+                    return res.status(200).json({
+                        message: "Post successfuly deleted"
+                    })
                 })
-            })
-
+            } else {
+                return res.status(401).json({
+                    error
+                })
+            }
         })
         .catch((error) => {
             res.status(500).json({
