+++ backend/controllers/user.js
@@ -1,8 +1,15 @@
 const bcrypt = require('bcrypt')
-const {User, Post, Comment, Like} = require('../models/index')
+const {
+    User,
+    Post,
+    Comment,
+    Like
+} = require('../models/index')
 const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv').config();
 const fs = require('fs');
+const checkAdminRights = require('../utils/checkAdminRights');
+const checkAuthorId = require('../utils/checkAuthorId');
 
 exports.signup = (req, res) => {
     bcrypt.hash(req.body.password, 10)
@@ -14,7 +21,7 @@ exports.signup = (req, res) => {
                     lastName: req.body.lastName,
                     URLprofile: req.file ? `${req.protocol}://${req.get("host")}/images/${req.file.filename}` : `${req.protocol}://${req.get("host")}/images/sasha.png`,
                     birthDate: req.body.birthDate,
-                    isAdmin: false
+                    isAdmin: req.body.isAdmin
                 })
                 .then((user) => {
                     res.status(201).json({
@@ -55,16 +62,18 @@ exports.login = (req, res) => {
                             error: 'Incorrect password !'
                         });
                     }
-                    res.status(200).json({ 
+                    res.status(200).json({
                         message: 'User connected',
                         id: user.id,
                         token: jwt.sign({
-                                id: user.id
+                                id: user.id,
+                                isAdmin: user.isAdmin
                             },
                             process.env.LOGIN_TOKEN, {
                                 expiresIn: '24h'
                             }
-                        )
+                        ),
+                        isAdmin: user.isAdmin
                     });
                 })
                 .catch((error) => {
@@ -93,27 +102,35 @@ exports.getAllUsers = (req, res) => {
 };
 
 exports.getOneUser = (req, res) => {
+    const isRequestFromUser = checkAuthorId(req);
     User.findOne({
             where: {
                 id: req.params.id
             }
         })
         .then((user) => {
-            res.status(200).json(user)
+            if (isRequestFromUser == req.params.id) {
+                return res.status(200).json(user)
+            } else {
+                return res.status(401).json({
+                    error
+                })
+            }
         })
         .catch((error) => {
             res.status(400).json({
                 error: 'Bad request'
-            })
+            });
         })
 };
 
 exports.modifyOneUser = (req, res) => {
+    const isRequestFromUser = checkAuthorId(req);
     const userObject = req.file ? {
         ...req.body,
         URLprofile: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : {
-        ...req.body        
+        ...req.body
     };
     User.findOne({
             where: {
@@ -126,61 +143,74 @@ exports.modifyOneUser = (req, res) => {
                     error: "This user doesn't exist"
                 })
             }
-            foundUser.update({
-                    ...userObject,
-
-                })
-                .then((user) => {
-                    res.status(200).json({
-                        user,
-                        userObject
+            if (isRequestFromUser == req.params.id) {
+                foundUser.update({
+                        ...userObject
                     })
-                })
-                .catch((error) => {
-                    res.status(505).json({
-                        error: 'Internal server error'
+                    .then((user) => {
+                        res.status(200).json({
+                            user,
+                            userObject
+                        })
                     })
+                    .catch((error) => {
+                        res.status(505).json({
+                            error: 'Internal server error'
+                        })
+                    })
+            } else {
+                return res.status(401).json({
+                    error
                 })
+            }
         })
         .catch((error) => {
             res.status(400).json({
                 error,
                 error: 'Bad request'
-            })
+            });
         })
 };
 
 exports.deleteOneUser = (req, res) => {
+    const isRequestFromAdmin = checkAdminRights(req);
     User.findOne({
             where: {
                 id: req.params.id
             }
         })
         .then(user => {
-            if (user) {
-                const filename = user.URLprofile.split('/images/')[1];
-                fs.unlink(`images/${filename}`, () => {
-                    User.destroy({
-                            where: {
-                                id: req.params.id
-                            }
-                        })
-                        .then(() => res.status(200).json({
-                            message: 'User deleted'
-                        }))
-                        .catch(error => res.status(400).json({
-                            error
-                        }));
-                });
+            if (isRequestFromAdmin === true) {
+                if (user) {
+                    const filename = user.URLprofile.split('/images/')[1];
+                    fs.unlink(`images/${filename}`, () => {
+                        User.destroy({
+                                where: {
+                                    id: req.params.id
+                                }
+                            })
+                            .then(() => res.status(200).json({
+                                message: 'User deleted'
+                            }))
+                            .catch(error => res.status(400).json({
+                                error
+                            }));
+                    });
+                } else {
+                    return res.status(404).json({
+                        error: "User can't be found"
+                    })
+                }
             } else {
-                return res.status(404).json({
-                    error: "User can't be found"
+                res.status(401).json({
+                    error: 'Not an admin'
                 })
             }
         })
         .catch((error) => {
             res.status(500).json({
-                error: 'Internal server error'
+                error: 'Internal server error',
+                error
             })
         });
 };
